refactor(audit): resolve structural finding in bench/parsing (#5517)
